Struct reactive_graph_graph::instances::entities::entity_instance::EntityInstance

pub struct EntityInstance {
    pub ty: EntityTypeId,
    pub id: Uuid,
    pub description: String,
    pub properties: PropertyInstances,
    pub extensions: Extensions,
}

Entity instances represents an typed object which contains properties.

The entity type defines the properties (name, data type and socket type).

In contrast to the entity type the entity instance stores values in it’s properties.

Fields
ty: EntityTypeId

The type definition of the entity type.

id: Uuid

The unique identifier of the entity instance.

description: String

The description of the entity instance.

properties: PropertyInstances

The properties of then entity instance.

Each property is represented by it’s name (String) and it’s value. The value is a representation of a JSON. Therefore the value can be boolean, number, string, array or an object. For more information about the data types please look at https://docs.serde.rs/serde_json/value/enum.Value.html

extensions: Extensions

Entity instance specific extensions.

Implementations

impl EntityInstance

pub fn builder() -> EntityInstanceBuilder<((), (), (), (), ())>

Create a builder for building EntityInstance. On the builder, call .ty(...), .id(...)(optional), .description(...)(optional), .properties(...)(optional), .extensions(...)(optional) to set the values of the fields. Finally, call .build() to create the instance of EntityInstance.

impl EntityInstance

pub fn new<T: Into<EntityTypeId>, P: Into<PropertyInstances>>( ty: T, id: Uuid, properties: P, ) -> EntityInstance

Constructs a new entity instance with the given type.

pub fn new_from_type<N: Into<String>, T: Into<String>, P: Into<PropertyInstances>>( namespace: N, type_name: T, id: Uuid, properties: P, ) -> EntityInstance

Constructs a new entity instance with the given namespace, type_name, id and properties.

pub fn new_without_properties<T: Into<EntityTypeId>>( ty: T, id: Uuid, ) -> EntityInstance

Constructs a new entity instance with the given type and id but without properties.