Trait reactive_graph_graph::instances::entities::entity_instance_container::NamespacedTypeEntityInstanceContainer

pub trait NamespacedTypeEntityInstanceContainer<T, NamespacedTypeDoesNotExistError, AddEntityInstanceError, UpdateEntityInstanceError, RemoveEntityInstanceError> {
    // Required methods
    fn entity_instances(
        &self,
        ty: &T,
    ) -> Result<EntityInstances, NamespacedTypeDoesNotExistError>;
    fn has_entity_instance(
        &self,
        ty: &T,
        id: Uuid,
    ) -> Result<bool, NamespacedTypeDoesNotExistError>;
    fn add_entity_instance(
        &self,
        ty: &T,
        entity_instance: EntityInstance,
    ) -> Result<(), AddEntityInstanceError>;
    fn update_entity_instance(
        &self,
        ty: &T,
        id: Uuid,
        entity_instance: EntityInstance,
    ) -> Result<(Uuid, EntityInstance), UpdateEntityInstanceError>;
    fn remove_entity_instance(
        &self,
        ty: &T,
        id: Uuid,
    ) -> Result<Option<(Uuid, EntityInstance)>, RemoveEntityInstanceError>;
}

Collection of a type which contains entity instances.

Required Methods

fn entity_instances( &self, ty: &T, ) -> Result<EntityInstances, NamespacedTypeDoesNotExistError>

Returns the entity instances.

fn has_entity_instance( &self, ty: &T, id: Uuid, ) -> Result<bool, NamespacedTypeDoesNotExistError>

Returns true, if an entity instance with the given id exists.

fn add_entity_instance( &self, ty: &T, entity_instance: EntityInstance, ) -> Result<(), AddEntityInstanceError>

Adds the given entity instance.

fn update_entity_instance( &self, ty: &T, id: Uuid, entity_instance: EntityInstance, ) -> Result<(Uuid, EntityInstance), UpdateEntityInstanceError>

Updates the entity instance with the id of the given entity instance.

fn remove_entity_instance( &self, ty: &T, id: Uuid, ) -> Result<Option<(Uuid, EntityInstance)>, RemoveEntityInstanceError>

Removes the entity instance with the given id.