reactive_graph_graph::instances::properties::property_instance

Struct PropertyInstances

pub struct PropertyInstances(/* private fields */);

Container for property instances. PropertyInstances exposes all functionality from the underlying DashMap<String, Value>.

Implementations

impl PropertyInstances

pub fn new() -> Self

Creates an empty map of property instances.

pub fn new_from_property_types_with_defaults( property_types: &PropertyTypes, ) -> Self

Creates property instances from a property type and initializes the values with the default values of the data type of the property.

pub fn property<N: Into<String>, V: Into<Value>>( self, name: N, value: V, ) -> Self

Add or replace property instance with the given name. Consumes and returns self (builder style).

pub fn names(&self) -> PropertyNames

Returns all property names.

pub fn to_map(&self) -> BTreeMap<String, Value>

Returns a sorted list of property instances.