reactive_graph_graph::instances::properties::property_instance_accessor

Trait PropertyInstanceGetter

pub trait PropertyInstanceGetter {
    // Required methods
    fn get<S: Into<String>>(&self, property_name: S) -> Option<Value>;
    fn as_bool<S: Into<String>>(&self, property_name: S) -> Option<bool>;
    fn as_u64<S: Into<String>>(&self, property_name: S) -> Option<u64>;
    fn as_i64<S: Into<String>>(&self, property_name: S) -> Option<i64>;
    fn as_f64<S: Into<String>>(&self, property_name: S) -> Option<f64>;
    fn as_string<S: Into<String>>(&self, property_name: S) -> Option<String>;
    fn as_array<S: Into<String>>(&self, property_name: S) -> Option<Vec<Value>>;
    fn as_object<S: Into<String>>(
        &self,
        property_name: S,
    ) -> Option<Map<String, Value>>;
}

Required Methods

fn get<S: Into<String>>(&self, property_name: S) -> Option<Value>

Returns the json value of the given property by name

fn as_bool<S: Into<String>>(&self, property_name: S) -> Option<bool>

Returns the boolean value of the given property by name

fn as_u64<S: Into<String>>(&self, property_name: S) -> Option<u64>

Returns the u64 value of the given property by name

fn as_i64<S: Into<String>>(&self, property_name: S) -> Option<i64>

Returns the i64 value of the given property by name

fn as_f64<S: Into<String>>(&self, property_name: S) -> Option<f64>

Returns the f64 value of the given property by name

fn as_string<S: Into<String>>(&self, property_name: S) -> Option<String>

Returns the string value of the given property by name

fn as_array<S: Into<String>>(&self, property_name: S) -> Option<Vec<Value>>

Returns the string value of the given property by name

fn as_object<S: Into<String>>( &self, property_name: S, ) -> Option<Map<String, Value>>

Returns the string value of the given property by name

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.