reactive_graph_graph::instances::properties::property_instance_accessor
Trait PropertyInstanceSetter

pub trait PropertyInstanceSetter: PropertyInstanceGetter {
    // Required methods
    fn set_checked<S: Into<String>>(&self, property_name: S, value: Value);
    fn set<S: Into<String>>(&self, property_name: S, value: Value);
    fn set_no_propagate_checked<S: Into<String>>(
        &self,
        property_name: S,
        value: Value,
    );
    fn set_no_propagate<S: Into<String>>(&self, property_name: S, value: Value);
    fn mutability<S: Into<String>>(
        &self,
        property_name: S,
    ) -> Option<Mutability>;
    fn set_mutability<S: Into<String>>(
        &self,
        property_name: S,
        mutability: Mutability,
    );
}

Required Methods

fn set_checked<S: Into<String>>(&self, property_name: S, value: Value)

Sets the value of the given property by name if the property is mutable.

fn set<S: Into<String>>(&self, property_name: S, value: Value)

Sets the value of the given property by name

fn set_no_propagate_checked<S: Into<String>>( &self, property_name: S, value: Value, )

Sets the value of the given property by name if the property is mutable. Sends the value down the stream.

fn set_no_propagate<S: Into<String>>(&self, property_name: S, value: Value)

Sets the value of the given property by name. Sends the value down the stream.

fn mutability<S: Into<String>>(&self, property_name: S) -> Option<Mutability>

Returns the mutability of the property by name.

fn set_mutability<S: Into<String>>( &self, property_name: S, mutability: Mutability, )

Sets the mutability of the property by name.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.