reactive_graph_graph::instances::relations::relation_instance_container

Trait NamespacedTypeRelationInstanceContainer

pub trait NamespacedTypeRelationInstanceContainer<T, NamespacedTypeDoesNotExistError, AddRelationInstanceError, UpdateRelationInstanceError, RemoveRelationInstanceError> {
    // Required methods
    fn relation_instances(
        &self,
        ty: &T,
    ) -> Result<RelationInstances, NamespacedTypeDoesNotExistError>;
    fn has_relation_which_uses_entity_instance(
        &self,
        ty: &T,
        id: Uuid,
    ) -> Result<bool, NamespacedTypeDoesNotExistError>;
    fn has_relation_instance(
        &self,
        ty: &T,
        id: &RelationInstanceId,
    ) -> Result<bool, NamespacedTypeDoesNotExistError>;
    fn add_relation_instance(
        &self,
        ty: &T,
        relation_instance: RelationInstance,
    ) -> Result<(), AddRelationInstanceError>;
    fn update_relation_instance(
        &self,
        ty: &T,
        id: &RelationInstanceId,
        relation_instance: RelationInstance,
    ) -> Result<(RelationInstanceId, RelationInstance), UpdateRelationInstanceError>;
    fn remove_relation_instance(
        &self,
        ty: &T,
        id: &RelationInstanceId,
    ) -> Result<Option<(RelationInstanceId, RelationInstance)>, RemoveRelationInstanceError>;
}

Collection of a type which contains relation instances.

Required Methods

fn relation_instances( &self, ty: &T, ) -> Result<RelationInstances, NamespacedTypeDoesNotExistError>

Returns the relation instances.

fn has_relation_which_uses_entity_instance( &self, ty: &T, id: Uuid, ) -> Result<bool, NamespacedTypeDoesNotExistError>

Returns true, if a relation instance exists which uses an entity instance with the given id.

fn has_relation_instance( &self, ty: &T, id: &RelationInstanceId, ) -> Result<bool, NamespacedTypeDoesNotExistError>

Returns true, if a relation instance with the given id exists.

fn add_relation_instance( &self, ty: &T, relation_instance: RelationInstance, ) -> Result<(), AddRelationInstanceError>

Adds the given relation instance.

fn update_relation_instance( &self, ty: &T, id: &RelationInstanceId, relation_instance: RelationInstance, ) -> Result<(RelationInstanceId, RelationInstance), UpdateRelationInstanceError>

Updates the relation instance with the id of the given relation instance.

fn remove_relation_instance( &self, ty: &T, id: &RelationInstanceId, ) -> Result<Option<(RelationInstanceId, RelationInstance)>, RemoveRelationInstanceError>

Removes the relation instance with the given id.