Trait reactive_graph_graph::instances::relations::relation_instance_container::RelationInstanceContainer

pub trait RelationInstanceContainer {
    // Required methods
    fn relation_instances(&self) -> RelationInstances;
    fn has_relation_which_uses_entity_instance(&self, id: Uuid) -> bool;
    fn has_relation_instance(&self, id: &RelationInstanceId) -> bool;
    fn add_relation_instance(
        &self,
        relation_instance: RelationInstance,
    ) -> Result<(), AddRelationInstanceError>;
    fn update_relation_instance(
        &self,
        id: &RelationInstanceId,
        relation_instance: RelationInstance,
    ) -> Result<(RelationInstanceId, RelationInstance), UpdateRelationInstanceError>;
    fn remove_relation_instance(
        &self,
        id: &RelationInstanceId,
    ) -> Result<Option<(RelationInstanceId, RelationInstance)>, RemoveRelationInstanceError>;
}

Container for relation instances.

Required Methods

fn relation_instances(&self) -> RelationInstances

Returns the relation instances.

fn has_relation_which_uses_entity_instance(&self, id: Uuid) -> bool

Returns true, if a relation instance exists which uses an entity instance with the given id.

fn has_relation_instance(&self, id: &RelationInstanceId) -> bool

Returns true, if a relation instance with the given id exists.

fn add_relation_instance( &self, relation_instance: RelationInstance, ) -> Result<(), AddRelationInstanceError>

Adds the given relation instance.

fn update_relation_instance( &self, id: &RelationInstanceId, relation_instance: RelationInstance, ) -> Result<(RelationInstanceId, RelationInstance), UpdateRelationInstanceError>

Updates the relation instance with the given id with the given relation instance.

fn remove_relation_instance( &self, id: &RelationInstanceId, ) -> Result<Option<(RelationInstanceId, RelationInstance)>, RemoveRelationInstanceError>

Removes the relation instance with the given id.