Trait reactive_graph_graph::types::components::component_container::ComponentContainer

pub trait ComponentContainer {
    // Required methods
    fn get_components(&self) -> ComponentTypeIds;
    fn add_component(&self, ty: ComponentTypeId);
    fn add_component_with_properties(&self, component: &Component);
    fn remove_component(&self, ty: &ComponentTypeId);
    fn is_a(&self, ty: &ComponentTypeId) -> bool;
}

Reactive instance container for components.

Required Methods

fn get_components(&self) -> ComponentTypeIds

Returns the component types of the container.

fn add_component(&self, ty: ComponentTypeId)

Adds a component to the container.

fn add_component_with_properties(&self, component: &Component)

Adds a component to the container and initializes the reactive property instances.

fn remove_component(&self, ty: &ComponentTypeId)

Removes a component from the container.

fn is_a(&self, ty: &ComponentTypeId) -> bool

Returns true, if the reactive instance is composed with the given component.