reactive_graph_graph::types::components::component_type_container

Trait ComponentTypeIdContainer

pub trait ComponentTypeIdContainer {
    // Required methods
    fn is_a(&self, ty: &ComponentTypeId) -> bool;
    fn add_component<C: Into<ComponentTypeId>>(&self, ty: C) -> bool;
    fn add_components<C: Into<ComponentTypeIds>>(
        &mut self,
        components_to_add: C,
    );
    fn remove_component(&self, ty: &ComponentTypeId) -> Option<ComponentTypeId>;
    fn remove_components<C: Into<ComponentTypeIds>>(
        &mut self,
        components_to_remove: C,
    );

    // Provided methods
    fn is_any(&self, component_tys: &ComponentTypeIds) -> bool { ... }
    fn is_all(&self, component_tys: &ComponentTypeIds) -> bool { ... }
}

Required Methods

fn is_a(&self, ty: &ComponentTypeId) -> bool

Returns true, if the type is composed with a component of the given type.

fn add_component<C: Into<ComponentTypeId>>(&self, ty: C) -> bool

Adds the given component.

fn add_components<C: Into<ComponentTypeIds>>(&mut self, components_to_add: C)

Adds the given components.

fn remove_component(&self, ty: &ComponentTypeId) -> Option<ComponentTypeId>

Removes the component with the given type id.

fn remove_components<C: Into<ComponentTypeIds>>( &mut self, components_to_remove: C, )

Removes the given components.

Provided Methods

fn is_any(&self, component_tys: &ComponentTypeIds) -> bool

Returns true, if the type is composed with any of the given components.

fn is_all(&self, component_tys: &ComponentTypeIds) -> bool

Returns true, if the type is composed with every given components.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.