reactive_graph_graph::types::components::component_type_container
Trait NamespacedTypeComponentTypeIdContainer

pub trait NamespacedTypeComponentTypeIdContainer<T, AddComponentError, RemoveComponentError> {
    // Required methods
    fn get_by_having_component(&self, component_ty: &ComponentTypeId) -> Self;
    fn add_component(
        &self,
        ty: &T,
        component_ty: &ComponentTypeId,
    ) -> Result<(), AddComponentError>;
    fn remove_component(
        &self,
        ty: &T,
        component_ty: &ComponentTypeId,
    ) -> Result<ComponentTypeId, RemoveComponentError>;
}

Required Methods

fn get_by_having_component(&self, component_ty: &ComponentTypeId) -> Self

Returns all types having the specified component.

fn add_component( &self, ty: &T, component_ty: &ComponentTypeId, ) -> Result<(), AddComponentError>

Adds the component with the given component_name to the entity type with the given name.

fn remove_component( &self, ty: &T, component_ty: &ComponentTypeId, ) -> Result<ComponentTypeId, RemoveComponentError>

Remove the component with the given component_name from the entity type with the given name.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.