Struct reactive_graph_graph::types::entities::entity_type::EntityType

pub struct EntityType {
    pub ty: EntityTypeId,
    pub description: String,
    pub components: ComponentTypeIds,
    pub properties: PropertyTypes,
    pub extensions: Extensions,
}

Entity types defines the type of an entity instance.

Fields
ty: EntityTypeId

The type definition contains the namespace and the type name.

description: String

Textual description of the entity type.

components: ComponentTypeIds

The names of the components of the entity type.

properties: PropertyTypes

The properties which are defined by the entity type.

extensions: Extensions

Entity type specific extensions.

Implementations

impl EntityType

pub fn builder() -> EntityTypeBuilder<((), (), (), (), ())>

Create a builder for building EntityType. On the builder, call .ty(...), .description(...)(optional), .components(...)(optional), .properties(...)(optional), .extensions(...)(optional) to set the values of the fields. Finally, call .build() to create the instance of EntityType.

impl EntityType

pub fn new<T: Into<EntityTypeId>, S: Into<String>, C: Into<ComponentTypeIds>, P: Into<PropertyTypes>, E: Into<Extensions>>( ty: T, description: S, components: C, properties: P, extensions: E, ) -> EntityType

Constructs an entity type from the given namespaced type with the given description, components, properties and extensions.

pub fn new_from_type<N: Into<String>, T: Into<String>, D: Into<String>, C: Into<ComponentTypeIds>, P: Into<PropertyTypes>, E: Into<Extensions>>( namespace: N, type_name: T, description: D, components: C, properties: P, extensions: E, ) -> EntityType

pub fn builder_from_ty<T: Into<EntityTypeId>>( ty: T, ) -> EntityTypeBuilder<((EntityTypeId,), (), (), (), ())>