Trait reactive_graph_graph::types::extensions::extension_container::ExtensionContainer

pub trait ExtensionContainer {
    // Required methods
    fn has_own_extension(&self, ty: &ExtensionTypeId) -> bool;
    fn get_own_extension(&self, ty: &ExtensionTypeId) -> Option<Extension>;
    fn add_extension<E: Into<Extension>>(
        &self,
        extension: E
    ) -> Result<ExtensionTypeId, AddExtensionError>;
    fn update_extension<T: Into<ExtensionTypeId>, E: Into<Extension>>(
        &self,
        ty: T,
        extension: E
    ) -> Result<Extension, UpdateExtensionError>;
    fn remove_extension<T: Into<ExtensionTypeId>>(
        &self,
        ty: T
    ) -> Result<Extension, RemoveExtensionError>;
    fn merge_extensions<E: Into<Extensions>>(&mut self, extensions_to_merge: E);

    // Provided methods
    fn has_any_own_extensions(&self, tys: &Vec<ExtensionTypeId>) -> bool { ... }
    fn has_all_own_extensions(&self, tys: &Vec<ExtensionTypeId>) -> bool { ... }
}

Required Methods

fn has_own_extension(&self, ty: &ExtensionTypeId) -> bool

Returns true, if the container has an extension with the given type.

fn get_own_extension(&self, ty: &ExtensionTypeId) -> Option<Extension>

Returns the own extension with the given type. Doesn’t respect extensions from potential components.

fn add_extension<E: Into<Extension>>( &self, extension: E ) -> Result<ExtensionTypeId, AddExtensionError>

fn update_extension<T: Into<ExtensionTypeId>, E: Into<Extension>>( &self, ty: T, extension: E ) -> Result<Extension, UpdateExtensionError>

fn remove_extension<T: Into<ExtensionTypeId>>( &self, ty: T ) -> Result<Extension, RemoveExtensionError>

fn merge_extensions<E: Into<Extensions>>(&mut self, extensions_to_merge: E)

Merge the given extensions into the own extensions.

Provided Methods

fn has_any_own_extensions(&self, tys: &Vec<ExtensionTypeId>) -> bool

Returns true, if the type contains any of the given extensions.

fn has_all_own_extensions(&self, tys: &Vec<ExtensionTypeId>) -> bool

Returns true, if the type contains all given extensions.

Object Safety

This trait is not object safe.