Struct reactive_graph_graph::types::flows::flow_type::FlowType

pub struct FlowType {
    pub ty: FlowTypeId,
    pub description: String,
    pub wrapper_entity_instance: EntityInstance,
    pub entity_instances: EntityInstances,
    pub relation_instances: RelationInstances,
    pub variables: Variables,
    pub extensions: Extensions,
}

Flow types defines the type of an flow instance like a template for flow instances.

They contain entity instances and relation instances. The wrapper entity instance is mandatory and used for input and outputs.

Fields
ty: FlowTypeId

The type definition of the entity type.

description: String

Textual description of the flow type.

wrapper_entity_instance: EntityInstance

The wrapper entity instance.

entity_instances: EntityInstances

The entity instances which are contained in this flow.

By default, no entity instances are contained in this flow type.

relation_instances: RelationInstances

The relation instances which are contained in this flow.

By default, no relation instances are contained in this flow type.

variables: Variables

The variables. Variables will be replaced by instantiation of a flow instance.

By default, the flow type has no variables.

extensions: Extensions

Flow type specific extensions.

Implementations

impl FlowType

pub fn builder() -> FlowTypeBuilder<((), (), (), (), (), (), ())>

Create a builder for building FlowType. On the builder, call .ty(...), .description(...)(optional), .wrapper_entity_instance(...), .entity_instances(...)(optional), .relation_instances(...)(optional), .variables(...)(optional), .extensions(...)(optional) to set the values of the fields. Finally, call .build() to create the instance of FlowType.

impl FlowType

pub fn new<T: Into<FlowTypeId>, D: Into<String>, EI: Into<EntityInstances>, RI: Into<RelationInstances>, V: Into<PropertyTypes>, E: Into<Extensions>>( ty: T, description: D, wrapper_entity_instance: EntityInstance, entity_instances: EI, relation_instances: RI, variables: V, extensions: E, ) -> FlowType

pub fn new_from_type<N: Into<String>, T: Into<String>, D: Into<String>, EI: Into<EntityInstances>, RI: Into<RelationInstances>, V: Into<PropertyTypes>, E: Into<Extensions>>( namespace: N, type_name: T, description: D, wrapper_entity_instance: EntityInstance, entity_instances: EI, relation_instances: RI, variables: V, extensions: E, ) -> FlowType

pub fn id(&self) -> Uuid

pub fn wrapper_type(&self) -> EntityTypeId

Returns the entity type namespace of the flow type.

pub fn uses_entity_types(&self) -> EntityTypeIds

Returns the entity types which are used by the flow type.

pub fn uses_relation_types(&self) -> RelationTypeIds

Returns the relation types which are used by the flow type.