Struct reactive_graph_graph::types::properties::property_type::PropertyType

pub struct PropertyType {
    pub name: String,
    pub description: String,
    pub data_type: DataType,
    pub socket_type: SocketType,
    pub mutability: Mutability,
    pub extensions: Extensions,
}

Definition of a property. The definition contains the name of the property, the data type and the socket type.

Fields
name: String

The name of the property

description: String

The description of the property.

data_type: DataType

The data type of the property

socket_type: SocketType

Specifies the type of socket - either input socket or output socket or none

mutability: Mutability

Specifies if the property is mutable.

extensions: Extensions

Property specific extensions

Implementations

impl PropertyType

pub fn builder() -> PropertyTypeBuilder<((), (), (), (), (), ())>

Create a builder for building PropertyType. On the builder, call .name(...), .description(...)(optional), .data_type(...), .socket_type(...)(optional), .mutability(...)(optional), .extensions(...)(optional) to set the values of the fields. Finally, call .build() to create the instance of PropertyType.

impl PropertyType

pub fn new<S: Into<String>>(name: S, data_type: DataType) -> PropertyType

pub fn new_with_socket<S: Into<String>>( name: S, data_type: DataType, socket_type: SocketType, ) -> PropertyType

pub fn input<S: Into<String>>(name: S, data_type: DataType) -> PropertyType

pub fn output<S: Into<String>>(name: S, data_type: DataType) -> PropertyType

pub fn new_with_all<S: Into<String>>( name: S, description: S, data_type: DataType, socket_type: SocketType, mutability: Mutability, extensions: Extensions, ) -> PropertyType

pub fn bool<S: Into<String>>(name: S) -> PropertyType

pub fn bool_input<S: Into<String>>(name: S) -> PropertyType

pub fn bool_output<S: Into<String>>(name: S) -> PropertyType

pub fn number<S: Into<String>>(name: S) -> PropertyType

pub fn number_input<S: Into<String>>(name: S) -> PropertyType

pub fn number_output<S: Into<String>>(name: S) -> PropertyType

pub fn string<S: Into<String>>(name: S) -> PropertyType

pub fn string_input<S: Into<String>>(name: S) -> PropertyType

pub fn string_output<S: Into<String>>(name: S) -> PropertyType

pub fn array<S: Into<String>>(name: S) -> PropertyType

pub fn array_input<S: Into<String>>(name: S) -> PropertyType

pub fn array_output<S: Into<String>>(name: S) -> PropertyType

pub fn object<S: Into<String>>(name: S) -> PropertyType

pub fn object_input<S: Into<String>>(name: S) -> PropertyType

pub fn object_output<S: Into<String>>(name: S) -> PropertyType

pub fn has_extension(&self, ty: &ExtensionTypeId) -> bool

Returns true, if the property contains an extension with the given type.

pub fn get_extension(&self, ty: &ExtensionTypeId) -> Option<Extension>