Trait reactive_graph_graph::types::properties::property_type_container::NamespacedTypePropertyTypeContainer

pub trait NamespacedTypePropertyTypeContainer<T, AddPropertyError, UpdatePropertyError, RemovePropertyError, MergePropertiesError> {
    // Required methods
    fn add_property<P: Into<PropertyType>>(
        &self,
        ty: &T,
        property_type: P,
    ) -> Result<PropertyType, AddPropertyError>;
    fn update_property<N: Into<String>, P: Into<PropertyType>>(
        &self,
        ty: &T,
        property_name: N,
        property_type: P,
    ) -> Result<PropertyType, UpdatePropertyError>;
    fn remove_property<N: Into<String>>(
        &self,
        ty: &T,
        property_name: N,
    ) -> Result<PropertyType, RemovePropertyError>;
    fn merge_properties<P: Into<PropertyTypes>>(
        &self,
        ty: &T,
        properties_to_merge: P,
    ) -> Result<(), MergePropertiesError>;
}

Collection of a type which contains property types.

Required Methods

fn add_property<P: Into<PropertyType>>( &self, ty: &T, property_type: P, ) -> Result<PropertyType, AddPropertyError>

Adds a property to the given type.

fn update_property<N: Into<String>, P: Into<PropertyType>>( &self, ty: &T, property_name: N, property_type: P, ) -> Result<PropertyType, UpdatePropertyError>

Updates the property with the given name of the given type. It’s possible to rename the property by using another name in the new property than the provided property_name.

fn remove_property<N: Into<String>>( &self, ty: &T, property_name: N, ) -> Result<PropertyType, RemovePropertyError>

Remove the property with the given name from the given type.

fn merge_properties<P: Into<PropertyTypes>>( &self, ty: &T, properties_to_merge: P, ) -> Result<(), MergePropertiesError>

Merges the given properties into the properties of the given type.

Object Safety

This trait is not object safe.