reactive_graph_graph::types::properties::property_type_container
Trait PropertyTypeContainer

pub trait PropertyTypeContainer {
    // Required methods
    fn has_own_property<S: Into<String>>(&self, property_name: S) -> bool;
    fn get_own_property<S: Into<String>>(
        &self,
        property_name: S,
    ) -> Option<PropertyType>;
    fn add_property<S: Into<PropertyType>>(
        &self,
        property_type: S,
    ) -> Result<PropertyType, AddPropertyError>;
    fn update_property<N: Into<String>, S: Into<PropertyType>>(
        &self,
        property_name: N,
        property_type: S,
    ) -> Result<PropertyType, UpdatePropertyError>;
    fn remove_property<S: Into<String>>(
        &self,
        property_name: S,
    ) -> Result<PropertyType, RemovePropertyError>;
    fn merge_properties<P: Into<PropertyTypes>>(
        &mut self,
        properties_to_merge: P,
    );

    // Provided methods
    fn has_any_own_properties(&self, property_names: &[String]) -> bool { ... }
    fn has_all_own_properties(&self, property_names: &[String]) -> bool { ... }
}

A type which contains property types.

Required Methods

fn has_own_property<S: Into<String>>(&self, property_name: S) -> bool

Returns true, if the type contains an own property with the given name. Doesn’t respect properties from potential components.

fn get_own_property<S: Into<String>>( &self, property_name: S, ) -> Option<PropertyType>

Returns the own property with the given name. Doesn’t respect properties from potential components.

fn add_property<S: Into<PropertyType>>( &self, property_type: S, ) -> Result<PropertyType, AddPropertyError>

Adds the given property.

fn update_property<N: Into<String>, S: Into<PropertyType>>( &self, property_name: N, property_type: S, ) -> Result<PropertyType, UpdatePropertyError>

Updates the property with the given property_name. It’s possible to rename the property by using another name in the new property than the provided property_name.

fn remove_property<S: Into<String>>( &self, property_name: S, ) -> Result<PropertyType, RemovePropertyError>

Removes the property with the given name.

fn merge_properties<P: Into<PropertyTypes>>(&mut self, properties_to_merge: P)

Merges the given properties into the own properties.

Provided Methods

fn has_any_own_properties(&self, property_names: &[String]) -> bool

Returns true, if the type contains any of the given properties. Doesn’t respect properties from potential components.

fn has_all_own_properties(&self, property_names: &[String]) -> bool

Returns true, if the type contains all of the given properties. Doesn’t respect properties from potential components.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.