reactive_graph_graph::types::variables::variable_container

Trait NamespacedTypeVariablesContainer

pub trait NamespacedTypeVariablesContainer<T, AddVariableError, UpdateVariableError, RemoveVariableError, MergeVariablesError> {
    // Required methods
    fn add_variable<V: Into<Variable>>(
        &self,
        ty: &T,
        variable: V,
    ) -> Result<Variable, AddVariableError>;
    fn update_variable<N: Into<String>, V: Into<Variable>>(
        &self,
        ty: &T,
        variable_name: N,
        variable: V,
    ) -> Result<Variable, UpdateVariableError>;
    fn remove_variable<N: Into<String>>(
        &self,
        ty: &T,
        variable_name: N,
    ) -> Result<Variable, RemoveVariableError>;
    fn merge_variables<V: Into<Variables>>(
        &mut self,
        ty: &T,
        variables_to_merge: V,
    ) -> Result<(), MergeVariablesError>;
}

Collection of a type which contains variables.

Required Methods

fn add_variable<V: Into<Variable>>( &self, ty: &T, variable: V, ) -> Result<Variable, AddVariableError>

Adds a variable to the given type.

fn update_variable<N: Into<String>, V: Into<Variable>>( &self, ty: &T, variable_name: N, variable: V, ) -> Result<Variable, UpdateVariableError>

Updates the variable with the given name of the given type. It’s possible to rename the variable by using another name in the new variable than the provided variable_name.

fn remove_variable<N: Into<String>>( &self, ty: &T, variable_name: N, ) -> Result<Variable, RemoveVariableError>

Remove the variable with the given name from the given type.

fn merge_variables<V: Into<Variables>>( &mut self, ty: &T, variables_to_merge: V, ) -> Result<(), MergeVariablesError>

Merges the given variables with the variables of the given type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.