Trait reactive_graph_graph::types::variables::variable_container::VariablesContainer

pub trait VariablesContainer {
    // Required methods
    fn has_variable<S: Into<String>>(&self, variable_name: S) -> bool;
    fn get_variable<N: Into<String>>(
        &self,
        variable_name: N,
    ) -> Option<Variable>;
    fn add_variable<V: Into<Variable>>(
        &self,
        variable: V,
    ) -> Result<Variable, AddVariableError>;
    fn update_variable<N: Into<String>, V: Into<Variable>>(
        &self,
        variable_name: N,
        variable: V,
    ) -> Result<Variable, UpdateVariableError>;
    fn remove_variable<S: Into<String>>(
        &self,
        variable_name: S,
    ) -> Result<Variable, RemoveVariableError>;
    fn merge_variables<V: Into<Variables>>(&mut self, variables_to_merge: V);

    // Provided methods
    fn has_any_variables(&self, variable_names: &VariableNames) -> bool { ... }
    fn has_all_variables(&self, variable_names: &VariableNames) -> bool { ... }
}

A type which contains variables.

Required Methods

fn has_variable<S: Into<String>>(&self, variable_name: S) -> bool

Returns true, if the type contains an variables with the given variable_name.

fn get_variable<N: Into<String>>(&self, variable_name: N) -> Option<Variable>

Returns the variable with the given variable_name.

fn add_variable<V: Into<Variable>>( &self, variable: V, ) -> Result<Variable, AddVariableError>

Adds the given variable.

fn update_variable<N: Into<String>, V: Into<Variable>>( &self, variable_name: N, variable: V, ) -> Result<Variable, UpdateVariableError>

Updates the variable with the given variable_name. It’s possible to rename the variable by using another name in the new variable than the provided variable_name.

fn remove_variable<S: Into<String>>( &self, variable_name: S, ) -> Result<Variable, RemoveVariableError>

Removes the variable with the given variable_name.

fn merge_variables<V: Into<Variables>>(&mut self, variables_to_merge: V)

Merges the given variables with the existing variables.

Provided Methods

fn has_any_variables(&self, variable_names: &VariableNames) -> bool

Returns true, if the type contains any of the given variables.

fn has_all_variables(&self, variable_names: &VariableNames) -> bool

Returns true, if the type contains all of the given variables.

Object Safety

This trait is not object safe.